revert refactor changes: changed new Model and ModelManager names, put the old ones back into the model package and refactored
