+++ src/main/java/seedu/address/logic/commands/BatchCommand.java
@@ -4,7 +4,7 @@ package seedu.address.logic.commands;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_INGREDIENT_ICECREAM;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_INGREDIENT_TOPPING;
 
-import oldcode.Model;
+import seedu.address.Model;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.IngredientModel;
