+++ src/main/java/seedu/address/logic/commands/HelpCommand.java
@@ -1,6 +1,6 @@
 package seedu.address.logic.commands;
 
-import oldcode.Model;
+import seedu.address.Model;
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.commons.events.ui.ShowHelpRequestEvent;
 import seedu.address.logic.CommandHistory;
