+++ src/main/java/seedu/address/logic/commands/HistoryCommand.java
@@ -5,7 +5,7 @@ import static java.util.Objects.requireNonNull;
 import java.util.Collections;
 import java.util.List;
 
-import oldcode.Model;
+import seedu.address.Model;
 import seedu.address.logic.CommandHistory;
