+++ src/test/java/seedu/address/TestApp.java
@@ -6,8 +6,6 @@ import java.util.function.Supplier;
 
 import javafx.stage.Screen;
 import javafx.stage.Stage;
-import oldcode.Model;
-import oldcode.ModelManager;
 import seedu.address.commons.core.Config;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.exceptions.DataConversionException;
