+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import oldcode.Model;
+import seedu.address.Model;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
