+++ src/test/java/seedu/address/logic/commands/HelpCommandTest.java
@@ -7,8 +7,8 @@ import static seedu.address.logic.commands.HelpCommand.SHOWING_HELP_MESSAGE;
 import org.junit.Rule;
 import org.junit.Test;
 
-import oldcode.Model;
-import oldcode.ModelManager;
+import seedu.address.Model;
+import seedu.address.ModelManager;
 import seedu.address.commons.events.ui.ShowHelpRequestEvent;
 import seedu.address.logic.CommandHistory;
 import seedu.address.ui.testutil.EventsCollectorRule;
