+++ src/test/java/seedu/address/logic/commands/ListCommandTest.java
@@ -8,13 +8,13 @@ import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 import org.junit.Before;
 import org.junit.Test;
 
-import oldcode.Model;
-import oldcode.ModelManager;
+import seedu.address.Model;
+import seedu.address.ModelManager;
 import seedu.address.logic.CommandHistory;
 import seedu.address.model.UserPrefs;
 
 /**
- * Contains integration tests (interaction with the Model) and unit tests for ListCommand.
+ * Contains integration tests (interaction with the DrinkModel) and unit tests for ListCommand.
  */
 public class ListCommandTest {
