+++ src/test/java/seedu/address/logic/commands/RedoCommandTest.java
@@ -8,8 +8,8 @@ import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
 import org.junit.Before;
 import org.junit.Test;
 
-import oldcode.Model;
-import oldcode.ModelManager;
+import seedu.address.Model;
+import seedu.address.ModelManager;
 import seedu.address.logic.CommandHistory;
 import seedu.address.model.UserPrefs;
