+++ src/test/java/seedu/address/testutil/TestUtil.java
@@ -5,7 +5,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-import oldcode.Model;
+import seedu.address.Model;
 import seedu.address.commons.core.index.Index;
 import seedu.address.model.person.Person;
